Move start-up related code to dedicated module
